+++ x-pack/plugins/aiops/server/routes/log_rate_analysis/define_route.ts
@@ -69,7 +69,7 @@ const PROGRESS_STEP_GROUPING = 0.1;
 const PROGRESS_STEP_HISTOGRAMS = 0.1;
 const PROGRESS_STEP_HISTOGRAMS_GROUPS = 0.1;
 
-export const defineLogRateAnalysisRoute = (
+export const defineRoute = (
   router: IRouter<DataRequestHandlerContext>,
   license: AiopsLicense,
   logger: Logger,
